DWORD type is unsigned long, not int
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@6321 212acab6-be3b-0410-9dea-997c60f758d6

+++ Source/Platform.h
@@ -39,7 +39,7 @@
 // basic types
 typedef unsigned char BYTE, *PBYTE, *LPBYTE;
 typedef unsigned short WORD, *LPWORD;
-typedef unsigned int DWORD, *LPDWORD;
+typedef unsigned long DWORD, *LPDWORD;
 typedef short SHORT;
 typedef unsigned short USHORT;
 typedef unsigned int UINT;
